+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/api/RetrofitCreatorUtils.kt
@@ -1,13 +1,13 @@
 package com.rakuten.tech.mobile.miniapp.api
 
+import android.os.Build
 import androidx.annotation.VisibleForTesting
 import com.google.gson.GsonBuilder
 import com.rakuten.tech.mobile.miniapp.BuildConfig
-import com.rakuten.tech.mobile.sdkutils.RasSdkHeaders
-import com.rakuten.tech.mobile.sdkutils.okhttp.addHeaderInterceptor
 import okhttp3.CertificatePinner
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
+import okhttp3.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import java.lang.IllegalArgumentException
@@ -23,10 +23,12 @@ internal fun createRetrofitClient(
     baseUrl = baseUrl,
     pubKeyList = pubKeyList,
     headers = RasSdkHeaders(
-        appId = rasProjectId,
-        subscriptionKey = subscriptionKey,
-        sdkName = "MiniApp",
-        sdkVersion = BuildConfig.VERSION_NAME
+        appId = "ras-app-id" to rasProjectId,
+        subscriptionKey = "apiKey" to "ras-$subscriptionKey",
+        sdkName = "ras-sdk-name" to "MiniApp",
+        sdkVersion = "ras-sdk-version" to BuildConfig.VERSION_NAME,
+        deviceModel = "ras-device-model" to Build.MODEL,
+        deviceOs = "ras-os-version" to Build.VERSION.RELEASE
     )
 )
 
@@ -85,4 +87,56 @@ internal fun extractBaseUrl(baseUrl: String): String {
     } catch (e: MalformedURLException) {
         ""
     }
+}
+
+/**
+ * Adds an Interceptor to OkHttp's okhttp3.OkHttpClient which will add the provided headers to all requests.
+ * @param headers the headers to be added to all requests.
+ */
+private fun OkHttpClient.Builder.addHeaderInterceptor(
+    vararg headers: Pair<String, String>
+): OkHttpClient.Builder = this.addNetworkInterceptor(HeaderInterceptor(headers))
+
+private class HeaderInterceptor constructor(
+    private val headers: Array<out Pair<String, String>>
+) : Interceptor {
+
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val requestBuilder = chain.request().newBuilder()
+        headers.forEach { header ->
+            requestBuilder.addHeader(header.first, header.second)
+        }
+
+        return chain.proceed(
+            requestBuilder.build()
+        )
+    }
+}
+
+/**
+ * Standard headers that should be sent with all requests to RAS.
+ */
+internal class RasSdkHeaders(
+    private val appId: Pair<String, String>,
+    private val subscriptionKey: Pair<String, String>,
+    private val sdkName: Pair<String, String>,
+    private val sdkVersion: Pair<String, String>,
+    private val deviceModel: Pair<String, String>,
+    private val deviceOs: Pair<String, String>
+) {
+
+    /**
+     * Returns the RAS headers as an array of [Pair]'s.
+     *
+     * @return array of [Pair] objects with [Pair.first] holding the header name
+     * and [Pair.second] holding the header value.
+     */
+    fun asArray() = arrayOf(
+        subscriptionKey,
+        appId,
+        sdkName,
+        sdkVersion,
+        deviceModel,
+        deviceOs
+    )
 }