+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebView.kt
@@ -43,7 +43,8 @@ internal open class MiniAppWebView(
         context,
         miniAppInfo,
         miniAppCustomPermissionCache,
-        miniAppFileChooser
+        miniAppFileChooser,
+        miniAppMessageBridge
     ),
     val queryParams: String,
     val ratDispatcher: MessageBridgeRatDispatcher,
@@ -159,7 +160,6 @@ internal open class MiniAppWebView(
     fun destroyView() {
         stopLoading()
         defaultFileDownloader.cleanup()
-        secureStorageDispatcher.cleanupSecureStorage()
         destroy()
     }