+++ CHANGELOG.md
@@ -4,9 +4,20 @@
 **SDK**
 - **Feature:** Added support for base64 `data:` URIs to the File Download feature in `MiniAppFileDownloader`.
 - **Feature:** Added secure storage support for storing, getting and removing data for MiniApps safely. HostApp can clear secured data using `MiniApp.clearSecureStorage`.
+- **Feature:** Added `MiniAppMessageBridge.miniAppShouldClose()` function which would help the host app to check if any alert need to be displayed before closing the MiniApp.
+- **Feature:** Added `MiniAppTooManyRequestsError` which will be thrown from SDK if any API from platform sends `429` status code.
 
 **Sample App**
 - **Feature:** Added a `Clear All` button to Settings/QA to remove all secure storages.
+- **Feature:** Added support to display error popup when there is `MiniAppTooManyRequestsError`.
+
+### 4.1.2 (2022-06-20)
+**SDK**
+- **Fix:** Removed `//` from API calls to generate a valid URL in `MetaDataApi` and `ManifestApi`.
+
+### 4.1.1 (2022-05-16)
+**SDK**
+- **Remove:** Removed `io.github.rakutentech.sdkutils` dependency and adopted the related changes to MiniApp SDK.
 
 ### 4.1.0 (2022-04-11)
 **SDK**